+++ common/build.gradle
@@ -1,0 +1,36 @@
+plugins {
+	id 'java-library'
+}
+
+def generatedDir = "src/main/generated"
+
+tasks.withType(JavaCompile).configureEach {
+	options.generatedSourceOutputDirectory.set(file(generatedDir))
+}
+
+idea {
+	module {
+		generatedSourceDirs += file(generatedDir)
+	}
+}
+
+clean {
+	delete file(generatedDir)
+}
+
+dependencies {
+	// spring
+	api 'org.springframework.boot:spring-boot-starter-webmvc'
+	api 'org.springframework.boot:spring-boot-starter-data-jpa'
+	api 'org.springframework.boot:spring-boot-starter-data-redis'
+	api 'org.springframework.boot:spring-boot-starter-validation'
+
+	// db
+	runtimeOnly 'org.postgresql:postgresql'
+
+	// querydsl
+	api "io.github.openfeign.querydsl:querydsl-jpa:${querydslVersion}"
+	annotationProcessor "io.github.openfeign.querydsl:querydsl-apt:${querydslVersion}:jakarta"
+	annotationProcessor "jakarta.persistence:jakarta.persistence-api"
+	annotationProcessor "jakarta.annotation:jakarta.annotation-api"
+}
