+++ common/src/main/java/bridge/bridge_bank/domain/ledger/entity/LedgerVoucher.java
