+++ common/src/main/java/bridge/bridge_bank/domain/ledger/repository/LedgerVoucherRepository.java
