+++ common/src/main/java/bridge/bridge_bank/domain/transfer_transaction_result/repository/TransferTransactionResultQueryRepository.java
