+++ core-server/build.gradle
@@ -1,0 +1,29 @@
+plugins {
+	id 'org.springframework.boot'
+}
+
+dependencies {
+	implementation project(':common')
+
+	// web
+	implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:3.0.2'
+
+	// observing
+	implementation 'org.springframework.boot:spring-boot-starter-actuator'
+	implementation 'net.ttddyy.observation:datasource-micrometer-spring-boot'
+	runtimeOnly 'io.micrometer:micrometer-registry-prometheus'
+
+	// batch
+	implementation 'org.springframework.boot:spring-boot-starter-batch'
+
+	// messaging
+	implementation 'org.springframework.boot:spring-boot-starter-kafka'
+
+	// test
+	testImplementation 'org.springframework.boot:spring-boot-starter-batch-test'
+	testImplementation 'org.springframework.boot:spring-boot-starter-data-jpa-test'
+	testImplementation 'org.springframework.boot:spring-boot-starter-data-redis-test'
+	testImplementation 'org.springframework.boot:spring-boot-starter-kafka-test'
+	testImplementation 'org.springframework.boot:spring-boot-starter-validation-test'
+	testImplementation 'org.springframework.boot:spring-boot-starter-webmvc-test'
+}
