+++ core-server/src/main/java/bridge/bridge_bank/domain/accounting_validation/reame.md
