+++ core-server/src/main/resources/application.yaml
