+++ reserve-once-transfer-worker/src/main/java/bridge/bridge_bank/domain/reserve_transfer_execution/ReserveOnceTransferExecutorConfig.java
@@ -1,0 +1,24 @@
+package bridge.bridge_bank.domain.reserve_transfer_execution;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
+
+import java.util.concurrent.Executor;
+import java.util.concurrent.ThreadPoolExecutor;
+
+@Configuration
+public class ReserveOnceTransferExecutorConfig {
+    @Bean(name = "asyncReserveTransferOnceExecutor")
+    public Executor asyncReserveOnceTransferExecutor() {
+        ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
+        executor.setCorePoolSize(8);
+        executor.setMaxPoolSize(8);
+        executor.setQueueCapacity(10000);
+        executor.setThreadNamePrefix("reserve-once-transfer-worker-");
+        executor.setKeepAliveSeconds(15);
+        executor.setRejectedExecutionHandler(new ThreadPoolExecutor.CallerRunsPolicy());
+        executor.initialize();
+        return executor;
+    }
+}
