+++ reserve-repeat-transfer-worker/src/main/resources/application.yaml
@@ -1,0 +1,20 @@
+spring:
+  application:
+    name: bridge-bank-reserve-repeat-worker
+  datasource:
+    url: jdbc:postgresql://${DB_HOST:localhost}:${DB_PORT:5432}/${DB_NAME:bridge_bank_db}
+    username: ${DB_USERNAME:admin}
+    password: ${DB_PASSWORD:1234}
+    driver-class-name: org.postgresql.Driver
+  data:
+    redis:
+      host: localhost
+      port: 6379
+  jpa:
+    hibernate:
+      ddl-auto: update
+    show-sql: true
+    properties:
+      hibernate:
+        format_sql: true
+        dialect: org.hibernate.dialect.PostgreSQLDialect
